pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "musicleague"
version = "2.0.0"
description = "Analysis toolkit for MusicLeague competitions"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "MusicLeague Team" }
]
keywords = ["music", "spotify", "analysis", "visualization", "competition"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "networkx>=3.0",
    "spotipy>=2.23.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
community = [
    "python-louvain>=0.16",  # For community detection in network analysis
]

[project.scripts]
musicleague-preprocess = "musicleague.scripts.preprocess:main"
musicleague-analyze = "musicleague.scripts.analyze:main"

[project.urls]
Homepage = "https://github.com/yourusername/musicleague"
Repository = "https://github.com/yourusername/musicleague"

[tool.hatch.build.targets.wheel]
packages = ["src/musicleague"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=musicleague"

